requirements.txt
python-dotenv==1.0.1
Requests==2.32.3
slack_bolt==1.22.0
